pyproject.toml
[tool.poetry]
name = "gramofzarcon"
version = "0.1.0"
description = ""
authors = ["sohail aslam <sohailaslam0707@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
Pillow = "^7.2.0"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
